src/responsive/requirements.txt
wxPython==4.1.0
Pillow==7.2.0
Pypubsub==4.0.3
Django==3.0.8
